% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{xrprof}
\alias{xrprof}
\alias{enable_no_sudo}
\title{Profile an External R Program}
\usage{
xrprof(file, pid = Sys.getpid(), frequency = 1, duration = 3600)

enable_no_sudo()
}
\arguments{
\item{file}{The file to write the samples to. See \code{\link[utils]{Rprof}}.}

\item{pid}{The process ID to target.}

\item{frequency}{The frequency to sample at, in hertz.}

\item{duration}{The length of time to profile, in seconds.}
}
\value{
A \code{\link[processx]{process}} instance.
}
\description{
\code{xrprof()} is similar to \code{\link[utils]{Rprof}} and produces the same
output format, but can be used to profile other running R programs instead.
}
